requirements.txt
python-telegram-bot==21.6
google-genai==1.10.0
groq==0.11.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
apscheduler==3.10.4
pydantic-settings==2.6.1
python-dotenv==1.0.1
